added some "return False"
I have problem when I use some Classes from other PY-script. When it fails I got lots of error in my PY-script. I tried to add some "return False" and now it seems to work better.
Is this the correct way of using "return False" or will it make some damage for the rest of the code?